requirements.txt
Pillow
stegano